@xiang90 2018-02-07
@mitake 

The commit routine is also probably the cause of what you observed https://github.com/coreos/etcd/pull/9199. 

Basically when you make the range size larger, it blocks the flush longer. Thus the writes batch is larger, and write can be more effective. But it is probably not what we want?