@xiang90 2018-02-08
I formed some new ideas to solve this issue + the problem #9199 tries to solve... so probably wait for my follow up commit :)