@xiang90 2018-02-08
/cc @heyitsanthony can you take another look? change to what you suggested.