@gyuho 2018-02-09
Approach looks good. I will try benchmarking this code path by issuing two overlapping readers contending on `backend.readTx`. If @mitake can cross-check on this, it would be great.