@hexfusion 2018-02-09
@gyuho would you mind outlining this process in a little detail?



> On Feb 8, 2018, at 7:07 PM, Gyuho Lee <notifications@github.com> wrote:
> 
> Approach looks good. I will try benchmarking this code path by issuing two overlapping readers contending on backend.readTx. If @mitake can cross-check on this, it would be great.
> 
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub, or mute the thread.
>