@gyuho 2018-02-09
@xiang90 Hard to benchmark exact code path with Go `Benchmark` since concurrent reads and commit for-loops are serialized on lock contentions. So spawned only two readers and manually overlapped commit code path.

```go
func BenchmarkBackendConcurrentRead(b *testing.B) {
	backend, tmppath := NewTmpBackend(10*time.Millisecond, 100000)
	defer backend.Close()
	defer os.Remove(tmppath)

	backend.batchTx.UnsafeCreateBucket([]byte("key"))

	// generate pending writes to "*batchTxBuffered"
	for i := 0; i < 100000; i++ {
		v := make([]byte, 128)
		rand.Read(v)
		backend.batchTx.Lock()
		backend.batchTx.UnsafePut([]byte("key"), []byte(fmt.Sprintf("foo%10d", i)), v)
		backend.batchTx.Unlock()
	}
	b.ResetTimer()

	k1, k2 := []byte(fmt.Sprintf("foo%10d", 0)), []byte(fmt.Sprintf("foo%10d", 100000))

	// simulate contention on "readTx.RWMutex":
	// one from periodic commit in batch transaction run for-loop
	// (without https://github.com/coreos/etcd/pull/9296)
	// the other to hold "readTx.RWMutex" with long read transaction
	readFunc := func() {
		now := time.Now()
		backend.ReadTx().Lock()
		backend.ReadTx().UnsafeRange([]byte("key"), k1, k2, 0)
		backend.ReadTx().Unlock()
		fmt.Println("read:", time.Since(now))
	}

	readN := 2
	for i := 0; i < 100; i++ {
		fmt.Println(i)
		var wg sync.WaitGroup
		wg.Add(readN * 2)
		for j := 0; j < readN; j++ {
			go func() {
				defer wg.Done()
				readFunc()
			}()
		}
		for j := 0; j < readN; j++ {
			go func() {
				defer wg.Done()
				now := time.Now()
				// backend.batchTx.Commit()
				if backend.batchTx.safePending() != 0 {
					backend.batchTx.Commit()
				}
				fmt.Println("commit:", time.Since(now))
			}()
		}
		wg.Wait()
	}
}
```

Before

```
commit: 16.725µs
read: 40.190658ms
read: 45.816537ms
commit: 45.839484ms

total 5.262s
```

After

```
commit: 323ns
commit: 152ns
read: 34.165254ms
read: 35.007701ms

total 3.689s
```

When pending writes are 0, this PR only holds `(*batchTxBuffered).Mutex` and does not block `ReadTx`, which seems to speed up 100 range requests with 2 concurrent readers.

LGTM. Thanks!